requirements.txt
llama-index-core==0.14.10
llama-index-readers-file==0.5.5
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-groq==0.4.1

torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
--extra-index-url https://download.pytorch.org/whl/cpu

sentence-transformers==3.0.1
numpy==1.26.4

python-dotenv==1.2.1
gradio==5.7.1
